Support for guilds with multiple characters of the same name in the same guild but on different realms.
Wow now supports having multiple characters with the same name but on
different realms on joint realms.  This update adds realms to being used
as a filter criteria for identifying character uniqueness.  It updates a
unique key in the DB and also updates the character claim screen to list
out realms in the options.